Updated ROLE enum values , Added new routes for ambassadors and system administrators.

+++ database/migrations/2025_07_09_153049_remove_biography_from_creators_table.php
@@ -11,9 +11,12 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('creators', function (Blueprint $table) {
-            $table->dropColumn('biography');
-        });
+        // check if column exists
+        if (Schema::hasColumn('creators', 'biography')) {
+            Schema::table('creators', function (Blueprint $table) {
+                $table->dropColumn('biography');
+            });
+        }
     }
 
     /**
